+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Make `MemoryRef.free()` synchronous (...).
+
 ## [0.1.0] - 2023-10-13
 
 ### Added